Seperate MetricsManager
